Make the response to search/beacons/:location consistent
