+++ content-service/server/database/beacons.js
@@ -1,22 +1,9 @@
 const { shortIdToNum, numToShortId, } = require('../utils/shortid');
 const HttpError = require('../express/httperror');
-const SHORT_URL = require('../../config.json').short_url;
-const nodeUrl = require('url');
-const resolveUrl = nodeUrl.resolve;
-const nodePath = require('path');
-
-const PARSED_SHORT_URL = Object.freeze(nodeUrl.parse(SHORT_URL));
+const utils = require('./utils');
 
 module.exports = function(knex) {
   const st = require('knex-postgis')(knex);
-
-  function selectBeacon() {
-    return knex('beacon')
-      .select(
-        'id', 'channel_name', 'canonical_url',
-        'call_to_action', 'extra_metadata',
-        st.asGeoJSON('location'), 'is_virtual');
-  }
 
   function createNewBeacon(channel, beaconData) {
     console.log('creating new beacon');
@@ -41,37 +28,6 @@ module.exports = function(knex) {
       });
   }
 
-  function mapDatabaseResponseToApiResponse(beacon) {
-    const location = JSON.parse(beacon.location);
-    const short = numToShortId(beacon.id);
-    return {
-      id: short,
-      short_url: resolveUrl(SHORT_URL, short),
-      channel: beacon.channel_name,
-      url: beacon.canonical_url,
-      call_to_action: JSON.parse(beacon.call_to_action),
-      extra_metadata: JSON.parse(beacon.extra_metadata),
-      location:  {
-        latitude: location.coordinates[1],
-        longitude: location.coordinates[0],
-      },
-      is_virtual: beacon.is_virtual,
-    };
-  }
-
-  function shortUrlToId(url) {
-    const parsedUrl = nodeUrl.parse(url);
-    const isShortUrl =
-      parsedUrl.host === PARSED_SHORT_URL.host &&
-      parsedUrl.protocol === PARSED_SHORT_URL.protocol;
-
-    if (!isShortUrl) {
-      return false;
-    }
-
-    return nodePath.basename(parsedUrl.pathname);
-  }
-
   function searchUrls(urls) {
     if (!Array.isArray(urls)) {
       throw new HttpError(400, 'Request body must be an array', 'EINVAL');
@@ -91,7 +47,7 @@ module.exports = function(knex) {
 
     const searchIds = urls
       .map((url) => {
-        const shortId = shortUrlToId(url);
+        const shortId = utils.shortUrlToId(url);
         // Not a magnet short URL, so add to the general search for urls
         if (!shortId) {
           searchUrls.push(url);
@@ -102,7 +58,7 @@ module.exports = function(knex) {
       .filter(shortId => !!shortId)
       .map(shortIdToNum);
 
-    let queryBuilder = selectBeacon();
+    let queryBuilder = utils.selectBeacons(knex);
 
     if (searchIds.length) {
       queryBuilder = queryBuilder.whereIn('id', searchIds);
@@ -113,7 +69,7 @@ module.exports = function(knex) {
     return queryBuilder
      .orWhereRaw('canonical_url = ANY(?::text[])', [searchUrls])
      .then((response) => {
-       return response.map(mapDatabaseResponseToApiResponse);
+       return response.map(utils.mapDatabaseResponseToApiResponse);
      });
   }
 
@@ -122,10 +78,10 @@ module.exports = function(knex) {
       throw new HttpError(400, 'Must specify channel name in request');
     }
 
-    return selectBeacon()
+    return utils.selectBeacons(knex)
       .where('channel_name', channelName)
       .then((response) => {
-        return response.map(mapDatabaseResponseToApiResponse);
+        return response.map(utils.mapDatabaseResponseToApiResponse);
       });
   }
 
@@ -134,10 +90,10 @@ module.exports = function(knex) {
       .filter(shortId => !!shortId)
       .map(shortIdToNum);
 
-    return selectBeacon()
+    return utils.selectBeacons(knex)
       .whereIn('id', ids)
       .then((response) => {
-        return response.map(mapDatabaseResponseToApiResponse);
+        return response.map(utils.mapDatabaseResponseToApiResponse);
       });
   }
 
@@ -150,7 +106,7 @@ module.exports = function(knex) {
       constraints['channel_name'] = channelName;
     }
 
-    return selectBeacon()
+    return utils.selectBeacons(knex)
       .where(constraints)
       .limit(1)
       .then((response) => {
@@ -160,7 +116,7 @@ module.exports = function(knex) {
 
         throw new HttpError(404, `Could not find beacon: ${slug}`, 'ENOTFOUND');
       })
-      .then(mapDatabaseResponseToApiResponse);
+      .then(utils.mapDatabaseResponseToApiResponse);
   }
 
   function updateBeacon(slug, patchData) {
@@ -187,7 +143,7 @@ module.exports = function(knex) {
       throw new HttpError(400, 'New beacon data does not include any properties available for update', 'EINVAL');
     }
 
-    return selectBeacon()
+    return knex('beacon')
       .where('id', shortIdToNum(slug))
       .update(updateObject)
       .returning(['channel_name', 'canonical_url', st.asGeoJSON('location')])
